\name{RESS-package}
\alias{RESS-package}
\alias{RESS}
\docType{package}
\title{
Integrates R and Essentia.
}
\description{
essQuery takes a single Essentia command and captures the output
in R, where you can save the output to a dataframe or stream it
directly into additional analysis.

read.udb takes a file containing any number of Essentia commands
and captures the output of the specified statements into R
dataframes.

read.udb can be used to stream multiple files directly into R with
a single command but essQuery is limited to one file per command.

}
\details{
\tabular{ll}{
Package: \tab RESS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-24\cr
License: \tab Same as Essentia. Free to use but do not make or distribute any copies.\cr
}
}
\author{
Ben Waxer, Data Scientist with Auriq Systems.

Maintainer: Who to complain to <bwaxer@auriq.com>
}
\keyword{ package }
